dev: merge wasm entries
