+++ scripts/build.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 cargo run --release --bin mitex
-cargo build --release --target wasm32-unknown-unknown -p mitex-typst
+cargo build --release --target wasm32-unknown-unknown --manifest-path ./crates/mitex-wasm/Cargo.toml --features typst-plugin
 cp target/wasm32-unknown-unknown/release/mitex_typst.wasm packages/mitex/mitex.wasm
 
-cd crates/mitex-web-wasm
-wasm-pack build --release
+cd crates/mitex-wasm
+wasm-pack build --release --features web
 cd ../..
